Move delete button to individual course screen
